LayoutTests/platform/wpe/editing/selection/select-out-of-editable-expected.txt

hello world
Dragging to the left: FAIL - expected 0 but got 7
Dragging to the left: FAIL - expected 11 but got 10
